+++ src/payload/collections/index.ts
@@ -1,8 +1,9 @@
+import { Categories } from "@/payload/collections/categories/schema";
 import { Media } from "@/payload/collections/media/schema";
 import { Pages } from "@/payload/collections/pages/schema";
 import { Posts } from "@/payload/collections/posts/schema";
 import { Users } from "@/payload/collections/users/schema";
 
-const collections = [Pages, Posts, Media, Users];
+const collections = [Pages, Posts, Categories, Media, Users];
 
 export { collections };